refactor: move telemetry code to separate file to remove circular dependency
